feat(stella-tui): port the SETTINGS tab's engine panel to the v2 SPEC 5 frame
The AGENTS pane of the SETTINGS tab rendered through views/engine.rs, a
grandfathered god file that drew a full-height bordered block with its own
title and legend on the border. On the SPEC 5 frame the tab strip above it
and the hint row below it already delimit the body, so that block spent two
columns and two rows repeating what the frame says.

The panel moves whole to crates/stella-tui/src/v2/engine_panel.rs, split
along three seams: tabs.rs is the strip's vocabulary and cycle order,
keys.rs the modal key map and the writes it makes into the working copy,
render.rs the five bands it draws. views/engine.rs and views/engine/tabs.rs
are deleted, and stella-tui loses a god file: the largest of the three new
modules is well under the flat ceiling, so the baseline entry retires.

The rendering itself is now v2 throughout: colours from
stella_tui_theme::token only, the selection lifted onto token::HL rather
than reversed, the active page in gold the way the deck's own tab row lights
one, and the modified marker on the strip's right edge where the block title
used to carry it. The model picker keeps a rounded border because it floats
over the rows and has to say where it ends.

Refs #4676
