+++ crates/stella-tui/src/views.rs
@@ -2,11 +2,11 @@
 //! `render(model: &WorkspaceModel, ui: &mut DeckUi, area: Rect, buf: &mut Buffer)`
 //! — a deterministic draw of the (model, ui) into a sub-area, recording any
 //! viewport metrics it needs for scroll clamping back onto `ui.metrics`.
-//! (`engine` and `tools` are the exceptions: they are the two config editors
-//! the SETTINGS tab ([`settings`]) hosts as the two panes of its ←/→ nav, not
-//! tab renderers of their own — each exposes `render_panel(ui, area, buf)`
-//! plus its own key handler, modal while that panel is focused. `installed`
-//! is a third: it
+//! (`tools` is the exception: it is one of the config editors the SETTINGS
+//! tab ([`settings`]) hosts as a pane of its ←/→ nav, not a tab renderer of
+//! its own — it exposes `render_panel(ui, area, buf)` plus its own key
+//! handler, modal while that panel is focused. The AGENTS pane beside it is
+//! [`crate::v2::engine_panel`]. `installed` is a second exception: it
 //! is the AGENTS tab's INSTALLED AGENTS pane, dispatched from
 //! [`agents::render`] rather than from the deck's tab match, and its
 //! `render(ui, now_ms, area, buf)` takes only the deck clock off the model rather than
@@ -35,7 +35,6 @@ pub mod approval;
 pub mod budget_card;
 pub(crate) mod cards;
 pub mod dispatch_card;
-pub mod engine;
 pub mod files;
 pub mod graph;
 pub mod installed;
